requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
Django==5.0.8
django-celery-beat==2.6.0
django-timezone-field==7.0
flower==2.0.1
gevent==24.2.1
greenlet==3.0.3
humanize==4.10.0
idna==3.7
kombu==5.4.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.8
requests==2.32.3
setuptools==72.2.0
six==1.16.0
sqlparse==0.5.1
tornado==6.4.1
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
zope.event==5.0
zope.interface==7.0.1
python_dotenv==1.0.1